egglog_experimental::prelude

Module sort

Structs

BaseValues

A registry for base value types and functions on them.

BigIntSort

BigRatSort

Rational numbers supporting these primitives:

BoolSort

Boxed

A newtype wrapper used to implement the BaseValue trait on types not defined in this crate.

ContainerValues

EqSort

ExecutionState

A handle on a database that may be in the process of running a rule.

F64Sort

64-bit floating point numbers supporting these primitives:

FunctionContainer

FunctionSort

I64Sort

Signed 64-bit integers supporting these primitives:

MapContainer

MapSort

A map from a key type to a value type supporting these primitives:

MultiSetContainer

MultiSetSort

OrderedFloat

A wrapper around floats providing implementations of Eq, Ord, and Hash.

PairContainer

PairSort

A pair of two values supporting these primitives:

ResolvedFunction

SetContainer

SetSort

StringSort

UnitSort

VecContainer

VecSort

Enums

ColumnTy

The type of a column (or QueryEntry): either an eq-sort / container id, or a base value (with the BaseValueId identifying which base type).

ResolvedFunctionId

Traits

ContainerValue

A trait implemented by container types.

Presort

Rebuilder

Custom functions used for tables that encode a bulk value-level rebuild of other tables.

Sort

A sort (type) in the e-graph that defines values in egglog. Sorts are user-extensible (e.g., prelude::BaseSort).

Functions

literal_sort

Type Aliases

F

MkSort

Q

S

Z